lua_state: Add integer overloads for pushOne
Otherwise, a small type mismatch (e.g. long long instead of int) will
cause ambiguity with e.g. the double overload.

+++ src/lua_state.cpp
@@ -77,7 +77,8 @@ void LuaState::pushOne(double value)
     lua_pushnumber(state(), value);
 }
 
-void LuaState::pushOne(int value)
+// NOLINTNEXTLINE(google-runtime-int)
+void LuaState::pushOne(long long value)
 {
     lua_pushinteger(state(), value);
 }
